fix(clang-tidy): fix ~590 clang-tidy warnings across ~60 files
Hybrid approach: disable 18 impractical checks (.clang-tidy) and fix
the remaining real issues found by clang-18 on ubuntu-24.04.

Disabled checks (with rationale in .clang-tidy):
- misc-include-cleaner (2380 IWYU violations, requires tooling)
- cppcoreguidelines-prefer-member-initializer (76, style-only)
- cppcoreguidelines-use-default-member-init (112, style-only)
- cppcoreguidelines-special-member-functions (67 Qt classes)
- performance-enum-size (53, may affect serialisation/ABI)
- cppcoreguidelines-pro-bounds-constant-array-index (29)
- cppcoreguidelines-avoid-const-or-ref-data-members (14)
- cppcoreguidelines-macro-usage (13, Qt macros unavoidable)
- readability-function-cognitive-complexity (11, deferred)
- cppcoreguidelines-avoid-do-while (10, valid pattern)
- cppcoreguidelines-pro-type-union-access (8, C API structs)
- misc-no-recursion (7, natural recursion)
- cppcoreguidelines-avoid-non-const-global-variables (7)
- cppcoreguidelines-avoid-goto (3, C-style error handling)
- readability-implicit-bool-conversion (3, Qt pattern)
- cppcoreguidelines-macro-to-enum (1, C/C++ boundary)
- bugprone-switch-missing-default-case (conflicts with -Wswitch-enum)

Fixed issues:
- cppcoreguidelines-init-variables: initialize Tox API error enums
  before passing by pointer in toxencrypt.cpp, toxoptions.cpp, v4l2.cpp,
  videoframe.cpp, stacktrace.cpp, privacyform.cpp, messageboxmanager.cpp
- bugprone-multi-level-implicit-pointer-conversion: videoframe.cpp av_freep
- bugprone-integer-division: screengrabberchooserrectitem.cpp (use 2.0),
  smileypack.cpp (use qsizetype throughout)
- bugprone-branch-clone: NOLINT on iaudiosink.h, history.cpp
- bugprone-assignment-in-if: rawdatabaseimpl.cpp, cameradevice.cpp,
  passwordedit.cpp (restructured)
- clang-analyzer-optin.cplusplus.VirtualCall: NOLINT on contentdialog.cpp,
  flowlayout.cpp, genericchatroomwidget.cpp
- performance-unnecessary-value-param: textcompose.h/cpp, aboutform.h/cpp,
  userinterfaceform.h/cpp, profileimporter.h/cpp, style.cpp,
  identicon.h/cpp, contentdialogmanager.h/cpp, conference.cpp,
  netcamview.cpp, autorun_xdg.cpp, ipc.h/cpp
- cppcoreguidelines-missing-std-forward: settings.cpp (std::forward
  in inGroup), stacktrace.h (NOLINT, &callback as void*)
- misc-const-correctness: corefile.cpp (fileCopy, fileNum), smileypack.cpp,
  genericchatform.cpp, v4l2.cpp
- modernize-use-auto: core.cpp, chatwidget.cpp, filesform.cpp
- performance-move-const-arg: conference.cpp (remove std::move from const&)
- concurrency-mt-unsafe: NOLINT on ipc.cpp, qrwidget.cpp, widget.cpp,
  posixsignalnotifier.cpp, v4l2.cpp
- cppcoreguidelines-pro-type-const-cast: NOLINT on coreav.cpp
- cert-dcl21-cpp: strongtype.h (const T postfix operator++)
- bugprone-integer-overflow: settingsserializer.cpp (100LL)
- performance-inefficient-algorithm: NOLINT on sessionchatlog.cpp
- cppcoreguidelines-no-malloc: NOLINT in stacktrace.cpp FreeDeleter
- performance-no-int-to-ptr: NOLINT in stacktrace.cpp libunwind usage
- bugprone-reserved-identifier: NOLINT on x11_display.h _XDisplay
- performance-move-constructor-init: chattextedit.cpp std::move(lm)
- bugprone-narrowing-conversions: corevideosource.cpp ptrdiff_t cast,
  textformatter.cpp qsizetype offset

+++ src/video/corevideosource.cpp
@@ -82,8 +82,8 @@ void CoreVideoSource::pushFrame(const vpx_image_t* vpxFrame)
         const int size = (i == 0) ? height : height / 2;
 
         for (int j = 0; j < size; ++j) {
-            uint8_t* dst = avFrame->data[i] + (dstStride * j);
-            uint8_t* src = vpxFrame->planes[i] + (srcStride * j);
+            uint8_t* dst = avFrame->data[i] + (static_cast<ptrdiff_t>(dstStride) * j);
+            uint8_t* src = vpxFrame->planes[i] + (static_cast<ptrdiff_t>(srcStride) * j);
             memcpy(dst, src, static_cast<size_t>(minStride));
         }
     }
